+++ src/jk_exceptionhelper/exception_helper.py
@@ -17,8 +17,15 @@ from .ExceptionObject import ExceptionObject
 # @param	bool ignoreJKLoggingFrames							(required) Flag to ignore stack frames from "jk_logging".
 # 																Such stacking frames don't help but distract from the essentials.
 #
-def analyseException(exception, ignoreJKTypingCheckFunctionSignatureFrames:bool = False, ignoreJKTestingAssertFrames:bool = False, ignoreJKLoggingFrames:bool = False) -> ExceptionObject:
-	return ExceptionObject.fromException(exception, ignoreJKTypingCheckFunctionSignatureFrames, ignoreJKTestingAssertFrames, ignoreJKLoggingFrames)
+def analyseException(
+		exception,
+		*,
+		ignoreJKTypingCheckFunctionSignatureFrames:bool = True,
+		ignoreJKTestingAssertFrames:bool = True,
+		ignoreJKLoggingFrames:bool = True,
+		ignoreJKPrettyPrintObjFrames:bool = True,
+	) -> ExceptionObject:
+	return ExceptionObject.fromException(exception, ignoreJKTypingCheckFunctionSignatureFrames, ignoreJKTestingAssertFrames, ignoreJKLoggingFrames, ignoreJKPrettyPrintObjFrames)
 #